+++ src/full/Agda/Auto/Auto.hs
@@ -332,10 +332,10 @@ auto ii rng argstr = do
                         case lookup mi riis of
                          Nothing ->
                           -- catchError
-                           (giveExpr Nothing mi expr >> return (Nothing, Nothing))
+                           (giveExpr WithoutForce Nothing mi expr >> return (Nothing, Nothing))
                            -- (const retry)
                            -- (\_ -> return (Nothing, Just ("Failed to give expr for side solution of " ++ show mi)))
-                         Just ii' -> do ae <- give ii' Nothing expr
+                         Just ii' -> do ae <- give WithoutForce ii' Nothing expr
                                         mv <- lookupMeta mi
                                         let scope = getMetaScope mv
                                         ce <- abstractToConcreteEnv (makeEnv scope) ae